Refactor schemas to consolidate storage and domain-related types, enhancing organization and type safety
