+++ lib/schemas/domain/seo.ts
