+++ lib/schemas/index.ts
@@ -1,8 +1,9 @@
-export * from "./dns";
-export * from "./export";
-export * from "./hosting";
-export * from "./http";
-export * from "./provider";
-export * from "./registration";
-export * from "./seo";
-export * from "./tls";
+export * from "./domain/certificates";
+export * from "./domain/dns";
+export * from "./domain/hosting";
+export * from "./domain/http";
+export * from "./domain/registration";
+export * from "./domain/seo";
+export * from "./internal/export";
+export * from "./internal/provider";
+export * from "./internal/storage";
