+++ lib/schemas/internal/export.ts
@@ -1,10 +1,10 @@
 import { z } from "zod";
-import { DnsRecordSchema, DnsResolverSchema } from "./dns";
-import { HostingSchema } from "./hosting";
-import { HttpHeadersSchema } from "./http";
-import { RegistrationSchema } from "./registration";
-import { SeoResponseSchema } from "./seo";
-import { CertificateSchema } from "./tls";
+import { CertificateSchema } from "../domain/certificates";
+import { DnsRecordSchema, DnsResolverSchema } from "../domain/dns";
+import { HostingSchema } from "../domain/hosting";
+import { HttpHeadersSchema } from "../domain/http";
+import { RegistrationSchema } from "../domain/registration";
+import { SeoResponseSchema } from "../domain/seo";
 
 export const DomainExportSchema = z.object({
   domain: z.string(),
