+++ lib/storage.ts
@@ -2,7 +2,7 @@ import "server-only";
 
 import { createHmac } from "node:crypto";
 import { UTApi, UTFile } from "uploadthing/server";
-import type { StorageKind } from "@/lib/schemas/storage";
+import type { StorageKind } from "@/lib/schemas";
 
 const ONE_WEEK_SECONDS = 7 * 24 * 60 * 60;
 const UPLOAD_MAX_ATTEMPTS = 3;
